Introduce Fluidbox to handle dynamic image modals
Adds Fluidbox 1.3.1a as a library for automatically managing image
modals.

At this point, any image in the DOM wrapped with an anchor element
will be activated as a possible modal when the page is ready. This
can be restricted in the future to a specific class of images or images
that are part of a container.

Additional configuration options are available through Fluidbox, but
the defaults work pretty well for our purposes. Do note that entirely
different images can be used in the modal via the anchor element